requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.14.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
scikit-image>=0.21.0
tqdm>=4.65.0
requests>=2.31.0
timm>=0.9.0
matplotlib>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0
scipy>=1.10.0
